# openapi.json
{
  "components": {
    "schemas": {
      "AnalysisCreated": {
        "properties": {
          "analysis_id": {
            "title": "Analysis Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "stream_url": {
            "title": "Stream Url",
            "type": "string"
          }
        },
        "required": [
          "analysis_id",
          "stream_url",
          "status"
        ],
        "title": "AnalysisCreated",
        "type": "object"
      },
      "AnalysisRequest": {
        "properties": {
          "query": {
            "maxLength": 800,
            "minLength": 3,
            "title": "Query",
            "type": "string"
          },
          "slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slug"
          }
        },
        "required": [
          "query"
        ],
        "title": "AnalysisRequest",
        "type": "object"
      },
      "AnalysisResult": {
        "properties": {
          "analysis_id": {
            "title": "Analysis Id",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "report": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MarketAnalysisReport"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "analysis_id",
          "status"
        ],
        "title": "AnalysisResult",
        "type": "object"
      },
...
      "CategoriesResponse": {
        "properties": {
          "categories": {
            "items": {
              "$ref": "#/components/schemas/InterestCategory"
            },
            "title": "Categories",
            "type": "array"
          }
        },
        "required": [
          "categories"
        ],
        "title": "CategoriesResponse",
        "type": "object"
      },
...
      "FeedbackRequest": {
        "properties": {
          "is_useful": {
            "title": "Is Useful",
            "type": "boolean"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "required": [
          "is_useful"
        ],
        "title": "FeedbackRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InterestsRequest": {
        "properties": {
          "categories": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "Categories",
            "type": "array"
          }
        },
        "required": [
          "categories"
        ],
        "title": "InterestsRequest",
        "type": "object"
      },
      "InterestsResponse": {
        "properties": {
          "interests": {
            "items": {
              "type": "string"
            },
            "title": "Interests",
            "type": "array"
          }
        },
        "required": [
          "interests"
        ],
        "title": "InterestsResponse",
        "type": "object"
      },
...
      "MeResponse": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "interests": {
            "items": {
              "type": "string"
            },
            "title": "Interests",
            "type": "array"
          },
          "is_grandfathered": {
            "title": "Is Grandfathered",
            "type": "boolean"
          },
          "is_invited": {
            "title": "Is Invited",
            "type": "boolean"
          },
          "onboarding_completed": {
            "title": "Onboarding Completed",
            "type": "boolean"
          },
          "usage": {
            "$ref": "#/components/schemas/UsageSummary"
          }
        },
        "required": [
          "id",
          "is_invited",
          "is_grandfathered",
          "onboarding_completed",
          "interests",
          "usage"
        ],
        "title": "MeResponse",
        "type": "object"
      },
...
      "ProblemResponse": {
        "description": "RFC 9457 problem+json.\n\nDeclared so the documented error shape matches what the API actually\nreturns; clients switch on `type`, never on `title` or `detail`.",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "title": "ProblemResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Causal explanations for Polymarket price moves.",
    "title": "PMIE API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/analyses": {
      "post": {
        "description": "Starts a run and returns immediately with a stream URL.\n\nA full analysis takes 60-120 seconds across four stages, so the work\nhappens in the background and progress arrives over the SSE endpoint.",
        "operationId": "create_analysis_v1_analyses_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              },
              "application/problem+json": {}
            },
            "description": "Not signed in, or the token failed verification"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              },
              "application/problem+json": {}
            },
            "description": "Signed in but not invited to the alpha"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              },
              "application/problem+json": {}
            },
            "description": "Rate limit exceeded"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              },
              "application/problem+json": {}
            },
            "description": "A dependency is unavailable"
          }
        },
        "summary": "Start an analysis"
      }
    },
    "/v1/analyses/{analysis_id}": {
      "get": {
        "description": "The finished report, for its owner or for a holder of a share link.\n\nThis route is listed in access.OPTIONAL_AUTH_ROUTES, so an anonymous\ncaller reaches this handler rather than being rejected outright \u2014 a\nshared report is meant to be readable without an account (UI_PRD 6.7).\nEverything else is decided here.\n\nOrder matters. Ownership is checked first so the owner never depends on a\ntoken; a share token is checked second; and only then does a signed-in\nstranger get 403 while an anonymous one gets 401. Collapsing those two\nwould either tell an anonymous caller that a report exists, or tell a\nsigned-in user to sign in again.",
        "operationId": "get_analysis_v1_analyses__analysis_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "title": "Analysis Id",
              "type": "string"
            }
          },
          {
            "description": "A link minted by the owner via POST .../share.",
            "in": "query",
            "name": "share_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "A link minted by the owner via POST .../share.",
              "title": "Share Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              },
              "application/problem+json": {}
            },
            "description": "Not signed in, or the token failed verification"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              },
              "application/problem+json": {}
            },
            "description": "Signed in but not invited to the alpha"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              },
              "application/problem+json": {}
            },
            "description": "No such analysis"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              },
              "application/problem+json": {}
            },
            "description": "A dependency is unavailable"
          }
        },
        "summary": "Get Analysis"
      }
    },
    "/v1/analyses/{analysis_id}/events": {
      "get": {
        "description": "Live stage events, for the owner only.\n\nNo share-token exception, deliberately: a link shares a finished report,\nnot a live view of someone else's run in progress.",
        "operationId": "stream_analysis_v1_analyses__analysis_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "title": "Analysis Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Server-sent events. One `stage_started` / `stage_completed` pair per pipeline stage (event_retrieval, signal_retrieval, news_retrieval, analysis), then a terminal `report` or `error` event. Each `data:` line is a JSON object `{stage, data}`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Analysis"
      }
    },
    "/v1/analyses/{analysis_id}/feedback": {
      "post": {
        "description": "Useful / not useful on a report.\n\nCloses the gap left by Phase 1: the API contract listed this endpoint but\nit was never built, so the UI had nothing to call.",
        "operationId": "submit_feedback_v1_analyses__analysis_id__feedback_post",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "title": "Analysis Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              },
              "application/problem+json": {}
            },
            "description": "Not signed in, or the token failed verification"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              },
              "application/problem+json": {}
            },
            "description": "Signed in but not invited to the alpha"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              },
              "application/problem+json": {}
            },
            "description": "A dependency is unavailable"
          }
        },
        "summary": "Rate a report"
      }
    },
...
    "/v1/interests/categories": {
      "get": {
        "description": "The selectable categories.\n\nServed from the database rather than hardcoded in the client so the list\nand its labels have one source of truth.",
        "operationId": "interest_categories_v1_interests_categories_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoriesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              },
              "application/problem+json": {}
            },
            "description": "A dependency is unavailable"
          }
        },
        "summary": "Selectable market categories"
      }
    },
...
    "/v1/me": {
      "get": {
        "description": "The caller's profile, interests and usage.\n\nOne call so the client can render the whole authenticated shell \u2014 header,\nusage indicator, onboarding state \u2014 without a waterfall of requests.",
        "operationId": "me_v1_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              },
              "application/problem+json": {}
            },
            "description": "Not signed in, or the token failed verification"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              },
              "application/problem+json": {}
            },
            "description": "Signed in but not invited to the alpha"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              },
              "application/problem+json": {}
            },
            "description": "A dependency is unavailable"
          }
        },
        "summary": "The signed-in user"
      }
    },
    "/v1/me/interests": {
      "put": {
        "operationId": "set_interests_v1_me_interests_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterestsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterestsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              },
              "application/problem+json": {}
            },
            "description": "Not signed in, or the token failed verification"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              },
              "application/problem+json": {}
            },
            "description": "Signed in but not invited to the alpha"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              },
              "application/problem+json": {}
            },
            "description": "Fewer than 3, more than 5, or unknown categories"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              },
              "application/problem+json": {}
            },
            "description": "A dependency is unavailable"
          }
        },
        "summary": "Choose 3-5 interest categories"
      }
    },
...
  }
}