# src/api/routes.py
from __future__ import annotations

import asyncio
import json
import logging

from fastapi import APIRouter, Depends, Query, Request
from fastapi.responses import StreamingResponse

from .access import get_current_user
from .accounts import FREE_MONTHLY_ANALYSES, AccountsError
from .auth import CurrentUser
from .config import describe_supabase_config
from .dependencies import require_invited
from .errors import ErrorType, PmieError
from .models import (
    AnalysisCreated,
    AnalysisRequest,
    AnalysisResult,
    CategoriesResponse,
    FeedbackRequest,
    HealthResponse,
    InterestsRequest,
    InterestsResponse,
    MeResponse,
    ProblemResponse,
    ReadyResponse,
    UsageSummary,
    extract_slug,
)
from .ratelimit import RateLimitExceeded

# Every failure path returns problem+json. Declaring it here means /docs shows
# the real error shape instead of FastAPI's default validation schema.
PROBLEM = {
    "model": ProblemResponse,
    "content": {"application/problem+json": {}},
}
_AUTH_ERRORS = {
    401: {"description": "Not signed in, or the token failed verification", **PROBLEM},
    403: {"description": "Signed in but not invited to the alpha", **PROBLEM},
    503: {"description": "A dependency is unavailable", **PROBLEM},
}

logger = logging.getLogger("cygnus.api.routes")

# Every route on this router requires a verified caller. Authorization used to
# be imperative — a handler was protected only if someone remembered to call
# current_user() in its body — which is how this endpoint's own GET and SSE
# stream shipped unauthenticated. Attaching it here inverts the default: a new
# route is closed unless it is named in access.PUBLIC_ROUTES.
router = APIRouter(prefix="/v1", dependencies=[Depends(get_current_user)])


@router.get("/health", response_model=HealthResponse, summary="Liveness")
async def health() -> dict:
    """Answers as soon as the process is up. Public: load balancers cannot
    authenticate."""
    return {"status": "ok", "service": "cygnus"}


@router.get("/ready", response_model=ReadyResponse, summary="Readiness")
async def ready(request: Request) -> dict:
    """Reports whether this process can do its job, not merely that it is up.

    Distinct from /health on purpose: a server that cannot persist reports
    answers /health perfectly well while silently discarding the only data
    the product's claims rest on.
    """
    check = getattr(request.app.state, "write_access", None)
    return {
        "status": "ready",
        "service": "cygnus",
        "checks": {
            "report_store_writable": ("unknown" if check is None else check.status),
            "detail": None if check is None else check.detail,
        },
    }


@router.post(
    "/analyses",
    status_code=201,
    response_model=AnalysisCreated,
    summary="Start an analysis",
    responses={
        **_AUTH_ERRORS,
        429: {"description": "Rate limit exceeded", **PROBLEM},
    },
)
async def create_analysis(
    body: AnalysisRequest,
    request: Request,
    user: CurrentUser = Depends(get_current_user),
):
    """Starts a run and returns immediately with a stream URL.

    A full analysis takes 60-120 seconds across four stages, so the work
    happens in the background and progress arrives over the SSE endpoint.
    """
    require_invited(request, user)

    # After the invite check, so an uninvited caller is told that rather than
    # being rate limited, and before any work starts — this endpoint is the
    # expensive one and the limit exists to protect the model budget.
    try:
        request.app.state.limiter.check(user.id)
    except RateLimitExceeded as exc:
        raise PmieError(
            ErrorType.RATE_LIMITED,
            f"Too many analyses. Try again in {exc.retry_after} seconds.",
            status=429,
        ) from exc

    registry = request.app.state.registry
    pipeline = request.app.state.pipeline
    slug = extract_slug(body.query, body.slug)
    if not slug:
        # Reject before any work starts. Without a slug the retrieval stages
        # have nothing to look up, and the analyst — correctly — produces an
        # "I had no market data" report. That costs four model calls and a
        # minute of the user's time to say what we could have said instantly,
        # and it stores a row that can never be scored.
        raise PmieError(
            ErrorType.INVALID_REQUEST,
            "Could not identify a Polymarket market in that request. Include "
            "the event slug or its full URL — for example "
            "'why is world-cup-winner moving?' or "
            "'https://polymarket.com/event/world-cup-winner'.",
            status=422,
        )

    record = registry.create(body.query, profile_id=user.id)

    # Keep a reference: a bare create_task can be garbage-collected mid-run,
    # which would silently strand the analysis in "running".
    task = asyncio.create_task(
        pipeline.run(record.analysis_id, body.query, slug, profile_id=user.id)
    )
    request.app.state.tasks.add(task)
    task.add_done_callback(request.app.state.tasks.discard)

    return AnalysisCreated(
        analysis_id=record.analysis_id,
        stream_url=f"/v1/analyses/{record.analysis_id}/events",
        status=record.status.value,
    )


def _share_token_grants(request, token: str | None, report_id: int | None) -> bool:
    """Whether this token is a live link to this specific report.

    report_id, not analysis_id: the in-process analysis record dies with the
    worker, while a share link has to keep meaning the same thing afterwards,
    so tokens are minted against the durable store row (see sharing.py).
    An unpersisted analysis has no report_id and can never be reached by a
    token — which is also why sharing_routes refuses to mint one.
    """
    if not token or report_id is None:
        return False
    sharing = getattr(request.app.state, "sharing", None)
    if sharing is None or not sharing.configured:
        return False
    return sharing.resolve(token) == report_id


@router.get(
    "/analyses/{analysis_id}",
    response_model=AnalysisResult,
    responses={
        **_AUTH_ERRORS,
        404: {"description": "No such analysis", **PROBLEM},
    },
)
async def get_analysis(
    analysis_id: str,
    request: Request,
    share_token: str | None = Query(
        None, description="A link minted by the owner via POST .../share."
    ),
    user: CurrentUser | None = Depends(get_current_user),
):
    """The finished report, for its owner or for a holder of a share link.

    This route is listed in access.OPTIONAL_AUTH_ROUTES, so an anonymous
    caller reaches this handler rather than being rejected outright — a
    shared report is meant to be readable without an account (UI_PRD 6.7).
    Everything else is decided here.

    Order matters. Ownership is checked first so the owner never depends on a
    token; a share token is checked second; and only then does a signed-in
    stranger get 403 while an anonymous one gets 401. Collapsing those two
    would either tell an anonymous caller that a report exists, or tell a
    signed-in user to sign in again.
    """
    record = request.app.state.registry.get(analysis_id)
    if record is None:
        raise PmieError(
            ErrorType.ANALYSIS_NOT_FOUND, f"no analysis {analysis_id}", status=404
        )

    if user is not None and record.profile_id == user.id:
        pass
    elif _share_token_grants(request, share_token, record.report_id):
        pass
    elif user is not None:
        raise PmieError(
            ErrorType.INVALID_REQUEST,
            "This analysis belongs to another account.",
            status=403,
        )
    else:
        # Deliberately identical to "not signed in": a revoked, expired or
        # invented token must not become an oracle for which analyses exist.
        raise PmieError(ErrorType.INVALID_REQUEST, "Sign in to continue.", status=401)

    return AnalysisResult(
        analysis_id=record.analysis_id,
        status=record.status.value,
        report=record.report,
        error=record.error,
    )


@router.get(
    "/analyses/{analysis_id}/events",
    # OpenAPI cannot describe an SSE frame sequence, but it can at least stop
    # claiming this returns JSON. Client generators key off the media type.
    response_class=StreamingResponse,
    responses={
        200: {
            "description": (
                "Server-sent events. One `stage_started` / `stage_completed` "
                "pair per pipeline stage (event_retrieval, signal_retrieval, "
                "news_retrieval, analysis), then a terminal `report` or "
                "`error` event. Each `data:` line is a JSON object "
                "`{stage, data}`."
            ),
            "content": {"text/event-stream": {"schema": {"type": "string"}}},
        }
    },
)
async def stream_analysis(
    analysis_id: str,
    request: Request,
    user: CurrentUser = Depends(get_current_user),
):
    """Live stage events, for the owner only.

    No share-token exception, deliberately: a link shares a finished report,
    not a live view of someone else's run in progress.
    """
    record = request.app.state.registry.get(analysis_id)
    if record is None:
        raise PmieError(
            ErrorType.ANALYSIS_NOT_FOUND, f"no analysis {analysis_id}", status=404
        )
    if record.profile_id != user.id:
        raise PmieError(
            ErrorType.INVALID_REQUEST,
            "This analysis belongs to another account.",
            status=403,
        )

    async def generator():
        while True:
            item = await record.queue.get()
            if item is None:
                break
            payload = json.dumps({"stage": item.stage, "data": item.data})
            yield f"event: {item.event}\ndata: {payload}\n\n"

    return StreamingResponse(
        generator(),
        media_type="text/event-stream",
        headers={"Cache-Control": "no-cache", "X-Accel-Buffering": "no"},
    )


@router.get(
    "/me",
    response_model=MeResponse,
    summary="The signed-in user",
    responses=_AUTH_ERRORS,
)
async def me(request: Request, user: CurrentUser = Depends(get_current_user)) -> dict:
    """The caller's profile, interests and usage.

    One call so the client can render the whole authenticated shell — header,
    usage indicator, onboarding state — without a waterfall of requests.
    """
    accounts = request.app.state.accounts
    if not accounts.configured:
        raise PmieError(
            ErrorType.INTERNAL_ERROR, "Accounts are not configured.", status=503
        )

    try:
        profile = accounts.get_profile(user.id)
        interests = accounts.get_interests(user.id) if profile else []
        used = accounts.monthly_usage(user.id) if profile else 0
    except AccountsError as exc:
        raise PmieError(
            ErrorType.INTERNAL_ERROR, "Could not load your account.", status=503
        ) from exc

    if profile is None:
        raise PmieError(
            ErrorType.INVALID_REQUEST, "No account found for this session.", status=401
        )

    return {
        "id": profile.id,
        "email": user.email,
        "display_name": profile.display_name,
        "is_invited": profile.is_invited,
        "is_grandfathered": profile.is_grandfathered,
        "onboarding_completed": profile.onboarding_completed_at is not None,
        "interests": interests,
        "usage": UsageSummary(
            analyses_this_month=used,
            free_monthly_allowance=FREE_MONTHLY_ANALYSES,
        ),
    }


@router.get(
    "/interests/categories",
    response_model=CategoriesResponse,
    summary="Selectable market categories",
    responses={503: _AUTH_ERRORS[503]},
)
async def interest_categories(request: Request) -> dict:
    """The selectable categories.

    Served from the database rather than hardcoded in the client so the list
    and its labels have one source of truth.
    """
    accounts = request.app.state.accounts
    if not accounts.configured:
        raise PmieError(
            ErrorType.INTERNAL_ERROR, "Accounts are not configured.", status=503
        )
    try:
        return {"categories": accounts.list_categories()}
    except AccountsError as exc:
        # Log the cause. Without this a misconfigured key produces a bare 503
        # and nothing actionable in the deployment's logs.
        logger.error(
            "could not load categories: %s | supabase config: %s",
            exc,
            describe_supabase_config(),
        )
        raise PmieError(
            ErrorType.INTERNAL_ERROR, "Could not load categories.", status=503
        ) from exc


@router.put(
    "/me/interests",
    response_model=InterestsResponse,
    summary="Choose 3-5 interest categories",
    responses={
        **_AUTH_ERRORS,
        422: {
            "description": "Fewer than 3, more than 5, or unknown categories",
            **PROBLEM,
        },
    },
)
async def set_interests(
    body: InterestsRequest,
    request: Request,
    user: CurrentUser = Depends(get_current_user),
) -> dict:
    accounts = request.app.state.accounts
    try:
        saved = accounts.set_interests(user.id, body.categories)
    except AccountsError as exc:
        raise PmieError(ErrorType.INVALID_REQUEST, str(exc), status=422) from exc
    return {"interests": saved}


@router.post(
    "/analyses/{analysis_id}/feedback",
    status_code=204,
    summary="Rate a report",
    responses=_AUTH_ERRORS,
)
async def submit_feedback(
    analysis_id: str,
    body: FeedbackRequest,
    request: Request,
    user: CurrentUser = Depends(get_current_user),
) -> None:
    """Useful / not useful on a report.

    Closes the gap left by Phase 1: the API contract listed this endpoint but
    it was never built, so the UI had nothing to call.
    """
    accounts = request.app.state.accounts
    try:
        accounts.save_feedback(user.id, analysis_id, body.is_useful, body.note)
    except AccountsError as exc:
        raise PmieError(
            ErrorType.INTERNAL_ERROR, "Could not save your feedback.", status=503
        ) from exc